todo: fix image upload bug

+++ src/main/java/com/shopwell/api/service/CartService.java
@@ -14,5 +14,5 @@ public interface CartService {
 
     List<CartItemVO> getCartItems(Long customerId);
 
-    BigDecimal calculateTotalPrice(Long customerId);
+    Double calculateTotalPrice(Long customerId);
 }
